# app/Http/Controllers/CarController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Support\Facades\DB;
use App\Models\Car;
use App\Models\CarImage;
use App\Models\ListingFee;
use App\Models\Platformpayment;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Storage;
use Illuminate\Support\Facades\Log;
use Illuminate\Support\Facades\Http;
use Exception;
use Illuminate\Support\Facades\Auth;



class CarController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/cars",
     *     summary="Get all cars",
     *     tags={"Car Management"},
     *     description="Retrieve a list of all available cars with their images",
     *     @OA\Response(
     *         response=200,
     *         description="List of cars retrieved successfully",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(ref="#/components/schemas/Car")
     *         )
     *     )
     * )
     */
    public function index()
    {
        return Car::with('images')->get();
    }

    /**
     * @OA\Get(
     *     path="/api/mycars",
     *     summary="Get user's cars",
     *     tags={"Car Management"},
     *     description="Retrieve all cars owned by the authenticated user",
     *     security={{"bearerAuth":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="User's cars retrieved successfully",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(ref="#/components/schemas/Car")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function myCars()
    {
            $userId = auth()->id();

            $cars = Car::with('images')
                ->where('owner_id', $userId)
                ->get();

            return response()->json($cars);
    }

    /**
     * @OA\Post(
     *     path="/api/cars",
     *     summary="Create a new car listing",
     *     tags={"Car Management"},
     *     description="Create a new car listing with images and payment processing for listing fees",
     *     security={{"bearerAuth":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\MediaType(
     *             mediaType="multipart/form-data",
     *             @OA\Schema(
     *                 required={"owner_id", "make", "model", "seating_capacity", "license_plate", "status", "price_per_day", "fuel_type", "transmission", "listing_type"},
     *                 @OA\Property(property="owner_id", type="integer", example=1, description="ID of the car owner"),
     *                 @OA\Property(property="make", type="string", example="Toyota", description="Car manufacturer"),
     *                 @OA\Property(property="model", type="string", example="Camry", description="Car model"),
     *                 @OA\Property(property="vin", type="string", example="1HGBH41JXMN109186", description="Vehicle identification number (optional)"),
     *                 @OA\Property(property="seating_capacity", type="integer", example=5, description="Number of seats"),
     *                 @OA\Property(property="license_plate", type="string", example="AB123CD", description="Car license plate"),
     *                 @OA\Property(property="status", type="string", example="Available", description="Car status"),
     *                 @OA\Property(property="price_per_day", type="number", format="float", example=50.00, description="Daily rental price"),
     *                 @OA\Property(property="fuel_type", type="string", enum={"Gasoline", "Diesel", "Electric", "Hybrid"}, example="Gasoline"),
     *                 @OA\Property(property="transmission", type="string", enum={"Manual", "Automatic"}, example="Automatic"),
     *                 @OA\Property(property="location_lat", type="number", format="float", example=9.0320, description="Latitude (optional)"),
     *                 @OA\Property(property="location_long", type="number", format="float", example=38.7615, description="Longitude (optional)"),
     *                 @OA\Property(property="pickup_location", type="string", example="Addis Ababa Airport", description="Pickup location (optional)"),
     *                 @OA\Property(property="return_location", type="string", example="Addis Ababa Airport", description="Return location (optional)"),
     *                 @OA\Property(property="listing_type", type="string", enum={"rent", "sell", "both"}, example="rent", description="Type of listing"),
     *                 @OA\Property(property="sell_price", type="number", format="float", example=25000.00, description="Sale price (if selling)"),
     *                 @OA\Property(property="is_negotiable", type="boolean", example=true, description="Whether price is negotiable"),
     *                 @OA\Property(property="mileage", type="integer", example=50000, description="Car mileage (optional)"),
     *                 @OA\Property(property="year", type="integer", example=2020, description="Manufacturing year (optional)"),
     *                 @OA\Property(property="condition", type="string", enum={"New", "Excellent", "Good", "Fair", "Poor"}, example="Good", description="Car condition (optional)"),
     *                 @OA\Property(property="images", type="array", @OA\Items(type="string", format="binary"), description="Car images (max 2MB each)")
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Car created successfully with payment redirect",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Car created. Redirect to Chapa for payment."),
     *             @OA\Property(property="checkout_url", type="string", example="https://checkout.chapa.co/checkout/payment/...")
     *         )
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Car created successfully without payment",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Car created successfully. No payment required."),
     *             @OA\Property(property="car", ref="#/components/schemas/Car")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     ),
     *     @OA\Response(
     *         response=500,
     *         description="Server error",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Error: Database transaction failed")
     *         )
     *     )
     * )
     */
    public function store(Request $request)
    {
        $user = Auth::user();
        if (Auth::check()) {
            logger('User ID for auditing: ' . Auth::id());
        } else {
            logger('No authenticated user found for auditing.');
        }

        $request->validate([
            'owner_id' => 'required|integer',
            'make'  =>  'required|string',
            'model' => 'required|string',
            'vin'   => 'nullable|string|unique:cars,vin',
            'seating_capacity' => 'required|integer',
            'license_plate' => 'required|string|unique:cars,license_plate',
            'status' => 'required|string',
            'price_per_day' => 'required|numeric',
            'fuel_type' => 'required|string',
            'transmission' => 'required|string',
            'location_lat' => 'nullable|numeric',
            'location_long' => 'nullable|numeric',
            'pickup_location' => 'nullable|string',
            'return_location' => 'nullable|string',
            'images.*' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',

            // Sale fields
            'listing_type' => 'required|in:rent,sell,both',
            'sell_price' => 'nullable|numeric',
            'is_negotiable' => 'nullable',
            'mileage' => 'nullable|integer',
            'year' => 'nullable|integer',
            'condition' => 'nullable|string',
        ]);

        DB::beginTransaction();
        try {
            $car = Car::create([
                'owner_id' => $request->owner_id,
                'make' => $request->make,
                'model' => $request->model,
                'vin' => $request->vin,
                'seating_capacity' => $request->seating_capacity,
                'license_plate' => $request->license_plate,
                'status' =>'pending',
                'price_per_day' => $request->price_per_day,
                'fuel_type' => $request->fuel_type,
                'transmission' => $request->transmission,
                'location_lat' => $request->location_lat ?? 8.9831,
                'location_long' => $request->location_long ?? 38.8101,
                'pickup_location' => $request->pickup_location,
                'return_location' => $request->return_location,

                // Sale fields
                'listing_type' => $request->listing_type,
                'sell_price' => $request->sell_price,
                'is_negotiable' => $request->is_negotiable ?? false,
                'mileage' => $request->mileage,
                'year' => $request->year,
                'condition' => $request->condition,
            ]);

            if ($request->hasFile('images')) {
                foreach ($request->file('images') as $image) {
                    $filename = time() . '_' . $image->getClientOriginalName();
                    $destinationPath = base_path('../public_html/car_images');
                    $image->move($destinationPath, $filename);

                    CarImage::create([
                        'car_id' => $car->id,
                        'image_path' => 'car_images/' . $filename,
                    ]);
                }
            }
        $fee = ListingFee::where('item_type', 'car')
            ->where('listing_type', $car->listing_type)
            ->orderByDesc('id') // get the latest active fee
            ->first();
        // Generate tx_ref and save payment
        $tx_ref = 'CARPOST-' . uniqid();
    
        if($fee && $fee->fee > 0 ) {
          Platformpayment::create([
            'item_id'=>$car->id,
            'item_type'=>'car',
            'amount' => $fee->fee,
            'currency' => $fee->currency,
            'payment_status' => 'pending',
            'payment_method' => 'chapa',
            'transaction_date' => now(),
            'tx_ref' => $tx_ref,
        ]);
    
    
        DB::commit();
        
        $returnUrl = $request->header('Platform') === 'mobile'
        ? url('/api/redirect/payment') . '?tx_ref=' . $tx_ref
        : env('Payment_FRONTEND_RETURN_URL') . '?tx_ref=' . $tx_ref;

        Log::info('Return URL', ['url' => $returnUrl]);

        
        $chapaData = [
            'amount' => $fee->fee,
            'currency' => 'ETB',
            'email' => auth()->user()->email,
            'first_name' => auth()->user()->first_name,
            'phone_number' => auth()->user()->phone,
            'tx_ref' => $tx_ref,
            'callback_url' => url('/api/chapa/listing-callback'),
            'return_url' => $returnUrl,
            'customization' => [
                'title' => 'Car Listing Fee',
                'description' => 'Payment to publish your car listing.',
            ],
        ];
     
        Log::info('Chapa Listing Payment Data', $chapaData);

        $response = Http::withToken(env('CHAPA_SECRET_KEY'))
            ->post(env('CHAPA_BASE_URL') . '/transaction/initialize', $chapaData);

        $data = $response->json();

        if ($response->successful() && isset($data['data']['checkout_url'])) {
            return response()->json([
                'message' => 'Car created. Redirect to Chapa for payment.',
                'checkout_url' => $data['data']['checkout_url'],
            ]);
        }

        return response()->json(['error' => 'Unable to redirect to Chapa.'], 500);

       }
        DB::commit();
        $car->load('images');
        return response()->json([
                'message' => 'Car created successfully. No payment required.',
                'car' => $car,
            ]);
    } catch (\Exception $e) {
            DB::rollBack();
            return response()->json(['message' => 'Error: ' . $e->getMessage()], 500);
        }
        
    
    }

    public function show($id)
    {
        return Car::with('images')->findOrFail($id);
    }

    public function update(Request $request, $id)
    {
        $user = Auth::user();
        $car = Car::findOrFail($id);

        $request->validate([
            'make' => 'sometimes|string',
            'model' => 'sometimes|string',
            'vin' => 'nullable|string|unique:cars,vin,' . $car->id,
            'seating_capacity' => 'sometimes|integer',
            'license_plate' => 'sometimes|string|unique:cars,license_plate,' . $car->id,
            'status' => 'sometimes|string',
            'price_per_day' => 'sometimes|numeric',
            'fuel_type' => 'sometimes|string',
            'transmission' => 'sometimes|string',
            'location_lat' => 'nullable|numeric',
            'location_long' => 'nullable|numeric',
            'pickup_location' => 'nullable|string',
            'return_location' => 'nullable|string',

            // Sale fields
            'listing_type' => 'sometimes|in:rent,sell,both',
            'sell_price' => 'nullable|numeric',
            'is_negotiable' => 'nullable|boolean',
            'mileage' => 'nullable|integer',
            'year' => 'nullable|integer',
            'condition' => 'nullable|string',
        ]);

        $car->update($request->all());

        return response()->json($car->fresh('images'));
    }

    public function destroy($id)
    {
        Car::destroy($id);
        return response()->noContent();
    }

    public function search(Request $request)
    {
        $query = Car::query();

        if ($request->has('make')) {
            $query->where('make', 'LIKE', '%' . $request->make . '%');
        }

        if ($request->has('model')) {
            $query->where('model', 'LIKE', '%' . $request->model . '%');
        }

        if ($request->has('seating_capacity')) {
            $query->where('seating_capacity', $request->seating_capacity);
        }

        if ($request->has('transmission')) {
            $query->where('transmission', $request->transmission);
        }

        if ($request->has('listing_type')) {
            $query->where('listing_type', $request->listing_type);
        }

        return response()->json($query->with('images')->get());
    }

    public function getCarImages($carId)
    {
        $car = Car::with('images')->find($carId);

        if (!$car) {
            return response()->json(['message' => 'Car not found'], 404);
        }

        return response()->json([
            'message' => 'Car images fetched successfully',
            'images' => $car->images,
        ]);
    }

    public function approveCar($id)
    {
        $car = Car::find($id);
        if (!$car) {
            return response()->json(['message' => 'Car not found'], 404);
        }

        $car->status = 'available';
        $car->save();

        return response()->json(['message' => 'Car listing approved']);
    }
    public function CarStatus(Request $request, $id)
        {
            $car = Car::find($id);
            if (!$car) {
                return response()->json(['message' => 'Car not found'], 404);
            }
            $request->validate([
                'status' => 'required|string',
            ]);
            $car->status = $request->status;
            $car->save();

            return response()->json(['message' => 'Car listing status updated successfully', 'car' => $car]);
        }


    public function rejectCar($id)
    {
        $car = Car::find($id);
        if (!$car) {
            return response()->json(['message' => 'Car not found'], 404);
        }

        $car->status = 'rejected';
        $car->save();

        return response()->json(['message' => 'Car listing rejected']);
    }

    public function blockCar($id)
    {
        $car = Car::find($id);
        if (!$car) {
            return response()->json(['message' => 'Car not found'], 404);
        }

        $car->status = 'blocked';
        $car->save();

        return response()->json(['message' => 'Car has been blocked']);
    }
}