# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Management System API",
        "description": "A comprehensive car rental management system with user authentication, vehicle management, booking system, payment processing, and review functionality. This system supports both car rentals and property (home) rentals.",
        "contact": {
            "name": "Car Rental System Support",
            "email": "support@carrentalsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Car Rental API Server"
        }
    ],
    "paths": {
...
        "/api/cars": {
            "get": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Get all cars",
                "description": "Retrieve a list of all available cars with their images",
                "operationId": "1fd4f1fdc0da258c5084bcdce4bbe693",
                "responses": {
                    "200": {
                        "description": "List of cars retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Create a new car listing",
                "description": "Create a new car listing with images and payment processing for listing fees",
                "operationId": "6f519ab2c9767d4d3b1be7cf8f1287ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "owner_id",
                                    "make",
                                    "model",
                                    "seating_capacity",
                                    "license_plate",
                                    "status",
                                    "price_per_day",
                                    "fuel_type",
                                    "transmission",
                                    "listing_type"
                                ],
                                "properties": {
                                    "owner_id": {
                                        "description": "ID of the car owner",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "make": {
                                        "description": "Car manufacturer",
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "model": {
                                        "description": "Car model",
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "vin": {
                                        "description": "Vehicle identification number (optional)",
                                        "type": "string",
                                        "example": "1HGBH41JXMN109186"
                                    },
                                    "seating_capacity": {
                                        "description": "Number of seats",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "license_plate": {
                                        "description": "Car license plate",
                                        "type": "string",
                                        "example": "AB123CD"
                                    },
                                    "status": {
                                        "description": "Car status",
                                        "type": "string",
                                        "example": "Available"
                                    },
                                    "price_per_day": {
                                        "description": "Daily rental price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "fuel_type": {
                                        "type": "string",
                                        "enum": [
                                            "Gasoline",
                                            "Diesel",
                                            "Electric",
                                            "Hybrid"
                                        ],
                                        "example": "Gasoline"
                                    },
                                    "transmission": {
                                        "type": "string",
                                        "enum": [
                                            "Manual",
                                            "Automatic"
                                        ],
                                        "example": "Automatic"
                                    },
                                    "location_lat": {
                                        "description": "Latitude (optional)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.032
                                    },
                                    "location_long": {
                                        "description": "Longitude (optional)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 38.7615
                                    },
                                    "pickup_location": {
                                        "description": "Pickup location (optional)",
                                        "type": "string",
                                        "example": "Addis Ababa Airport"
                                    },
                                    "return_location": {
                                        "description": "Return location (optional)",
                                        "type": "string",
                                        "example": "Addis Ababa Airport"
                                    },
                                    "listing_type": {
                                        "description": "Type of listing",
                                        "type": "string",
                                        "enum": [
                                            "rent",
                                            "sell",
                                            "both"
                                        ],
                                        "example": "rent"
                                    },
                                    "sell_price": {
                                        "description": "Sale price (if selling)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000
                                    },
                                    "is_negotiable": {
                                        "description": "Whether price is negotiable",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "mileage": {
                                        "description": "Car mileage (optional)",
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "year": {
                                        "description": "Manufacturing year (optional)",
                                        "type": "integer",
                                        "example": 2020
                                    },
                                    "condition": {
                                        "description": "Car condition (optional)",
                                        "type": "string",
                                        "enum": [
                                            "New",
                                            "Excellent",
                                            "Good",
                                            "Fair",
                                            "Poor"
                                        ],
                                        "example": "Good"
                                    },
                                    "images": {
                                        "description": "Car images (max 2MB each)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car created successfully with payment redirect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car created. Redirect to Chapa for payment."
                                        },
                                        "checkout_url": {
                                            "type": "string",
                                            "example": "https://checkout.chapa.co/checkout/payment/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Car created successfully without payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car created successfully. No payment required."
                                        },
                                        "car": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error: Database transaction failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mycars": {
            "get": {
                "tags": [
                    "Car Management"
                ],
                "summary": "Get user's cars",
                "description": "Retrieve all cars owned by the authenticated user",
                "operationId": "b6e06d1bd1d78b7f876c7567b40897e3",
                "responses": {
                    "200": {
                        "description": "User's cars retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "The field is required."
                            ]
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}